content_import: Fix crash in post handler.
The condition would call `count($j['item'])` if `$j['item']` was null,
which will throw a TypeError in php 8.0 and later. (It's been a warning
since php 7.2!) This breaks the import.

Also used `isset()` to check if the variable is set and not null before
calling count. Again, it avoids a warning in case the variable is not
set.


(cherry picked from commit 36b513115f0cbe4bc5d8c405de057ef824da43e7)